requirements.txt
alembic==1.16.2
astroid==3.3.10
autopep8==2.3.2
babel==2.17.0
backrefs==5.9
biopython==1.85
black==25.1.0
certifi==2025.6.15
cfgv==3.4.0
charset-normalizer==3.4.2
click==8.2.1
colorama==0.4.6
colorlog==6.9.0
contourpy==1.3.2
coverage==7.9.2
cycler==0.12.1
dill==0.4.0
distlib==0.3.9
filelock==3.18.0
fonttools==4.58.5
ghp-import==2.1.0
greenlet==3.2.3
identify==2.6.12
idna==3.10
iniconfig==2.1.0
isort==6.0.1
Jinja2==3.1.6
joblib==1.5.1
kiwisolver==1.4.8
lxml==5.4.0
Mako==1.3.10
Markdown==3.8.2
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.10.3
mccabe==0.7.0
mdurl==0.1.2
mergedeep==1.3.4
mkdocs==1.6.1
mkdocs-autorefs==1.4.2
mkdocs-gen-files==0.5.0
mkdocs-get-deps==0.2.0
mkdocs-material==9.6.15
mkdocs-material-extensions==1.3.1
multidict==6.6.3
multiprocess==0.70.18
mypy_extensions==1.1.0
narwhals==1.45.0
nodeenv==1.9.1
numpy==2.3.1
optuna==4.4.0
packaging==25.0
paginate==0.5.7
pandas==2.3.0
pathspec==0.12.1
pillow==11.3.0
platformdirs==4.3.8
plotly==6.2.0
pluggy==1.6.0
pre_commit==4.2.0
propcache==0.3.2
psutil==7.0.0
pycodestyle==2.14.0
Pygments==2.19.2
pylint==3.3.7
pymdown-extensions==10.16
pyparsing==3.2.3
pytest==8.4.1
pytest-cov==6.2.1
pytest-vcr==1.0.2
python-dateutil==2.9.0.post0
python-docx==1.2.0
pytz==2025.2
PyYAML==6.0.2
pyyaml_env_tag==1.1
requests==2.32.4
rich==14.0.0
ruff==0.12.2
SALib==1.5.1
scikit-learn==1.7.0
scipy==1.16.0
six==1.17.0
SQLAlchemy==2.0.41
tabulate==0.9.0
threadpoolctl==3.6.0
tomlkit==0.13.3
tqdm==4.67.1
typing_extensions==4.14.0
tzdata==2025.2
urllib3==2.5.0
vcrpy==7.0.0
virtualenv==20.31.2
watchdog==6.0.0
wrapt==1.17.2
yarl==1.20.1
python-dotenv==1.0.1
typer==0.9.0
